feat(maqueen): improve `fineTurnAdjustment()` with accelerometer and gyroscope
Also add some sound effects, and an option to pause and resume a run by
using the on-board buttons or the logo as a toggle (this was very useful
during debugging). Also also, adjust settings on MAX_FINE_ADJUSTMENT_ANGLE
and FINE_TURN_TIMOUT_MS to make better use of the new feature (I still
have not seen the timeout be reached unless the robot was physically
stuck on an obstacle).